+++ packages/components/modal/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tw-material/modal
 
+## 1.1.5
+
+### Patch Changes
+
+- update
+
 ## 1.1.4
 
 ### Patch Changes